breaking dataloader code into multiple files
